requirements.txt
aiohttp~=3.8.5
aiohttp-cors~=0.7.0
dateparser~=1.1.8
dnspython~=2.4.2
loggly-python-handler~=1.0.1
mailchimp-marketing~=3.0.80
mongomock~=4.1.2
Pillow~=10.0.1
pip-check-updates~=0.22.0
pyOpenSSL~=23.2.0
pymongo~=4.5.0
python-dateutil~=2.8.2
pytest~=7.4.2
pytz~=2023.3.post1
PyYAML~=6.0.1
requests~=2.31.0
requests-futures~=1.0.1
sendgrid~=6.10.0
simplejson~=3.19.1